+++ variables.tf
@@ -113,6 +113,60 @@ variable "ssh_hmac_sha2_512" {
   default     = true
 }
 
+variable "ssh_curve25519_sha256" {
+  description = "curve25519-sha256 kex algorithm."
+  type        = bool
+  default     = false
+}
+
+variable "ssh_curve25519_sha256_libssh" {
+  description = "curve25519-sha256 libssh.org kex algorithm."
+  type        = bool
+  default     = false
+}
+
+variable "ssh_dh1_sha1" {
+  description = "diffie-hellman-group1-sha1 kex algorithm."
+  type        = bool
+  default     = false
+}
+
+variable "ssh_dh14_sha1" {
+  description = "diffie-hellman-group14-sha1 kex algorithm."
+  type        = bool
+  default     = false
+}
+
+variable "ssh_dh14_sha256" {
+  description = "diffie-hellman-group14-sha256 kex algorithm."
+  type        = bool
+  default     = false
+}
+
+variable "ssh_dh16_sha512" {
+  description = "diffie-hellman-group16-sha512 kex algorithm."
+  type        = bool
+  default     = false
+}
+
+variable "ssh_ecdh_sha2_nistp256" {
+  description = "ecdh-sha2-nistp256 kex algorithm."
+  type        = bool
+  default     = false
+}
+
+variable "ssh_ecdh_sha2_nistp384" {
+  description = "ecdh-sha2-nistp384 kex algorithm."
+  type        = bool
+  default     = false
+}
+
+variable "ssh_ecdh_sha2_nistp521" {
+  description = "ecdh-sha2-nistp521 kex algorithm."
+  type        = bool
+  default     = false
+}
+
 variable "https_admin_state" {
   description = "HTTPS admin state."
   type        = bool