+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.1 (unreleased)
+
+- Support HTTPS TLS v1.3
+- Support SSH AES256-GCM cipher
+- Support SSH KEX Algorithms
+
 ## 0.1.0
 
 - Migrate to devnet provider